# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kosova Estate",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/careers": {
            "post": {
                "summary": "Submit a job application",
                "operationId": "b4af4f9088271e734af4300ad21c3065",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Applicant's name",
                                        "type": "string"
                                    },
                                    "surname": {
                                        "description": "Applicant's surname",
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "description": "Applicant's birthday",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "nationality": {
                                        "description": "Applicant's nationality",
                                        "type": "string"
                                    },
                                    "personal_number": {
                                        "description": "Applicant's personal number",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Applicant's email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "message": {
                                        "description": "Applicant's message",
                                        "type": "string"
                                    },
                                    "resume": {
                                        "description": "Applicant's resume",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Form submitted successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/job-applications": {
            "get": {
                "summary": "Get all job applications",
                "operationId": "aaabbf1a132d64b12b892078ad702733",
                "responses": {
                    "200": {
                        "description": "List of job applications"
                    },
                    "500": {
                        "description": "Unable to fetch job applications"
                    }
                }
            }
        },
        "/api/job-applications/{id}/approve": {
            "post": {
                "summary": "Approve a job application",
                "operationId": "13e900684c1d0a9540d310e56f764e28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application approved successfully"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                }
            }
        },
        "/api/job-applications/{id}/reject": {
            "post": {
                "summary": "Reject a job application",
                "operationId": "48dfd2b5941eda6518143814454b9ba5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application rejected successfully"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                }
            }
        },
        "/api/job-applications/{id}/undo": {
            "post": {
                "summary": "Undo job application status",
                "operationId": "8a3b5ac76a06e95fdfc388cf9ec32167",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application status undone successfully"
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "500": {
                        "description": "Unable to undo application status"
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "summary": "Get all clients",
                "operationId": "0b636ca85e574ff3010b44d6c621c537",
                "responses": {
                    "200": {
                        "description": "List of clients"
                    },
                    "500": {
                        "description": "Unable to fetch clients"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "put": {
                "summary": "Update a client",
                "operationId": "9ffd4acb0368eb6068e75cd7dd5d43ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name",
                                        "type": "string"
                                    },
                                    "username": {
                                        "description": "Username",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "Phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unable to update user"
                    }
                }
            },
            "delete": {
                "summary": "Delete a client",
                "operationId": "527e3694096ac0e9d71511de8b647c8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unable to delete user"
                    }
                }
            }
        },
...
    }
}